+++ src/main/scala/is/hail/rvd/UnpartitionedRVD.scala
@@ -50,24 +50,11 @@ class UnpartitionedRVD(val rowType: TStruct, val crdd: ContextRDD[RVDContext, Re
     new UnpartitionedRVD(rowType, crdd.sample(withReplacement, p, seed))
 
   def zipWithIndex(name: String): UnpartitionedRVD = {
-    val (newRowType, ins) = rowType.unsafeStructInsert(TInt64(), List(name))
-
-    val a = sparkContext.broadcast(countPerPartition().scanLeft(0L)(_ + _))
+    val (newRowType, newCRDD) = zipWithIndexCRDD(name)
 
     new UnpartitionedRVD(
-      rowType = newRowType.asInstanceOf[TStruct],
-      crdd = crdd.cmapPartitionsWithIndex({ (i, ctx, it) =>
-        val rv2 = RegionValue()
-        val rvb = ctx.rvb
-        var index = a.value(i)
-        it.zipWithIndex.map { case (rv, i) =>
-          rvb.start(newRowType)
-          ins(rv.region, rv.offset, rvb, () => rvb.addLong(index))
-          index += 1
-          rv2.set(rvb.region, rvb.end())
-          rv2
-        }
-      }, preservesPartitioning=true)
+      rowType = newRowType,
+      crdd = crdd
     )
   }
